refactor: added folders with css module files and reorganize project structure
