--- pequeña_prueba.txt
@@ -1,1 +1,0 @@
-Hola, soy Miguel y estoy realizando un pequeño push
