+++ src/App.jsx
@@ -1,28 +1,23 @@
-import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import './App.css';
-
-import Error from './components/Error';
-import HomePage from './view/HomePage';
-import Login from './view/Login';
-import Register from './view/Register';
-import ForgotPassword from './view/ForgotPassword';
-import LandingPage from './view/LandingPage';
-import NavBar from './components/NavBar';
-import Footer from './components/Footer';
-// import TestComponent from './components/TestComponent';
-import Page404 from './view/Page404';
-import ModProject from './view/ModProject';
-import Users from './components/User/Users';
-import AddProjectForm from './view/createForm';
-import useAuth0TokenHandler from './hooks/useAuth0TokenHandler';
 import { useAuth0 } from '@auth0/auth0-react';
-import Detail from './view/Detail';
+
+import Error from "./components/error/Error"
+import Footer from './components/footer/Footer';
+import NavBar from './components/nav-bar/NavBar';
+import HomePage from './view/home-page/HomePage';
+import LoginPage from './view/login-page/LoginPage';
+import RegisterPage from './view/register-page/RegisterPage';
+import ForgotPasswordPage from './view/forgot-password-page/ForgotPasswordPage';
+import LandingPage from './view/landing-page/LandingPage';
+import NotFoundPage from './view/not-found-page/NotFoundPage';
+import UpdateProjectPage from './view/update-project-page/UpdateProjectPage';
+import UsersPage from './view/users-page/UsersPage';
+import CreateProjectPage from './view/create-project-page/CreateFormPage';
+import useAuth0TokenHandler from './hooks/useAuth0TokenHandler';
+import ProjectDetailPage from './view/project-detail-page/ProjectDetailPage';
 
 function App() {
   const { isLoading } = useAuth0();
-  const { fetchError } = useSelector((state) => state.errors);
 
   // Asegúrate de que este hook se llame siempre en cada renderizado
   useAuth0TokenHandler();
@@ -32,19 +27,19 @@ function App() {
   return (
     <>
       <div>
-        {fetchError && <Error error={fetchError} />}
+        <Error />
         <NavBar />
         <Routes className="App">
           <Route path="/" element={<LandingPage />} />
-          <Route path="/create" element={<AddProjectForm />} />
+          <Route path="/create" element={<CreateProjectPage />} />
           <Route path="/home" element={<HomePage />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/register" element={<Register />} />
-          <Route path="/forgotPassword" element={<ForgotPassword />} />
-          <Route path="/modProject" element={<ModProject />} />
-          <Route path="/*" element={<Page404 />} />
-          <Route path="/users" element={<Users />} />
-		  <Route path="/project/detail/:id" element={<Detail />} />
+          <Route path="/login" element={<LoginPage />} />
+          <Route path="/register" element={<RegisterPage />} />
+          <Route path="/forgotPassword" element={<ForgotPasswordPage />} />
+          <Route path="/modProject" element={<UpdateProjectPage />} />
+          <Route path="/*" element={<NotFoundPage />} />
+          <Route path="/users" element={<UsersPage />} />
+          <Route path="/project/:id" element={<ProjectDetailPage />} />
         </Routes>
         <Footer />
       </div>
