--- src/components/logoutButton.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { useAuth0 } from '@auth0/auth0-react';
-import { useDispatch } from 'react-redux';
-import { logout as reduxLogout } from '../redux/actions';
-
-const LogoutButton = () => {
-  const { logout } = useAuth0();
-  const dispatch = useDispatch();
-
-  const handleLogout = () => {
-    dispatch(reduxLogout());
-    logout({ returnTo: window.location.origin });
-  };
-
-  return (
-    <button onClick={handleLogout} className="btn btn-secondary">
-      Salir
-    </button>
-  );
-};
-
-export default LogoutButton;
