+++ src/components/searchbar/Searchbar.module.css
@@ -1,0 +1,42 @@
+header{
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+  header .search-box{
+    flex: 1;
+    position:relative;
+    width: 100%;
+  }
+  
+  .search-box input{
+    width: 100%;
+    height: 40px;
+    border: none;
+    outline: none;
+    background: #dbdbdb;
+    border: none;
+    border-radius: 30px;
+    color: grey;
+    font-size: 16px;
+    text-align: center;
+    
+  }
+  
+  .search-box button{
+    cursor: pointer;
+    width: 50px;
+    height: 40px;
+    border-radius: 30px;
+    border: none;
+    position: absolute;
+    top: 0;
+    right: 0;
+    transform: scale(0.9);
+    background-color: aquamarine;
+  }
+  
+  .custom-checkbox{
+    width: "50px";
+    height: "50px";
+  }
